pyproject.toml
[project]
name = "zwift"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars (>=1.31.0,<2.0.0)",
    "fitparse (>=1.2.0,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "google-cloud-bigquery (>=3.35.0,<4.0.0)",
    "pandas-gbq (>=0.29.2,<0.30.0)",
    "dbt-bigquery (>=1.10.1,<2.0.0)",
    "streamlit (>=1.38.0,<2.0.0)",
    "plotly (>=5.24.0,<6.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "src" }]
[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

